Completed 2.14 (Optional) Making the camera follow the player
Games need to have the camera focus on the play area, so that the player does not miss out on any action. The
position of the camera depends on the type of game (First Person Shooters vs. 2D Platformers). The camera in
this case should follow the ”player object” that was defined earlier